refactor(wizard): #1544 stage 8 — extract create_course generateLessonPlanAndReturn
Stage 8 of 8, completing #1544. Lifts the ~160-LOC tail of create_course
(teachMethod backfill, instruction-spec sync, pedagogy assertions
#1545 mirror block, background `generateInstantCurriculum` kickoff,
first-call preview assembly, and the final WizardToolExec response
payload) into `tools/create_course/_lesson-plan.ts` exporting
`generateLessonPlanAndReturn(args)`.

The file name keeps the brief's terminology even though the actual
generateLessonPlan call was removed in an earlier change ("scheduler
handles pacing") — preserved for traceability with the issue's
recommended sequencing.

Static-regression test at `tests/lib/chat/wizard-tool-executor/pedagogy-assertions.test.ts`
updated: post-#1544 the pedagogy block lives in `_reuse-path.ts`
(reuse-branch) + `_lesson-plan.ts` (new-path branch). The test now
surveys orchestrator + both stage helpers as a combined source so a
future regression in either branch still fires the same #1545 drift
check.

Two prior-commit type tightenings reverted to preserve byte-equivalent
behaviour with the pre-extract `as any` casts: `existingGoals` in
`_new-config-merge.ts` and `finalFlowPhases` in `_new-path.ts`. Both
now carry `eslint-disable` comments matching the surrounding tolerance.

Net file shape:
- create_course.ts: 255 → 108 LOC (final orchestrator shape)
- new: _lesson-plan.ts (196 LOC)

Final tally vs pre-#1544 (1321-LOC monolith):
- create_course.ts: 1321 → 108 LOC (-92%)
- 10 new helper files under tools/create_course/, none over 400 LOC
  (_new-path.ts 399, _reuse-path.ts 393, _lesson-plan.ts 196,
  _enroll.ts 158, _new-config-merge.ts 153, _resolve-domain.ts 94,
  _reuse-config-merge.ts 79, _graph-guard.ts 55, _resolve-subject.ts 50,
  _context.ts 31)

Tests
- 12/12 dispatcher pins green (including the NEW new-path #607 pin from
  Stage 5)
- 161/161 tests/lib/chat suite green
- 0 new tsc errors introduced
- AC met: All 8 stages extracted, orchestrator simplified to a stage
  sequence, every helper under 400 LOC, new-path test pin landed.

Closes #1544
